wip: fmt, fix vec<enum>

+++ egui_field_editor/examples/shared/simple.rs
@@ -16,7 +16,7 @@ struct MyApp {
 	boolean: bool,
 	raw_string: &'static str,
 	#[inspect(slider(min = "-43.0", max = 125.0))]
-	float64: f32,
+	float64: f64,
 	#[inspect(name = "A proper field name")]
 	ugly_internal_field_name: u16,
 }
